+++ upkie/cpp/exceptions/FilterError.h
@@ -1,27 +1,24 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2022 Stéphane Caron
+// Copyright 2024 Inria
 
 #pragma once
 
 #include <string>
 
+#include "upkie/cpp/exceptions/UpkieError.h"
+
 namespace upkie::cpp::exceptions {
 
 //! Exception for filter errors
-class FilterError : public std::runtime_error {
+class FilterError : public UpkieError {
  public:
   /*! Constructor.
    *
    * \param[in] message Error message.
    */
   explicit FilterError(const std::string& message) throw()
-      : std::runtime_error(message.c_str()) {}
-
-  /*! Get error message.
-   *
-   * \return Error message.
-   */
-  const char* what() const throw() { return std::runtime_error::what(); }
+      : UpkieError(message) {}
 };
 
 }  // namespace upkie::cpp::exceptions